+++ apps/assets/src/components/asset/modals/transactions/hooks/useDeposit.ts
@@ -134,10 +134,8 @@ export const useDeposit = (useDepositProps: DepositProps) => {
       wallet.osmosisPubkey,
       keplrAddress,
       params,
-      chainIds.chainIdentifier.toUpperCase(),
       wallet.extensionName,
       prefix,
-      chainIds.chainId,
       chainIds.chainIdentifier
     );