+++ apps/governance/src/components/governance/proposals/ContainerProposals.tsx
@@ -39,6 +39,7 @@ const ContainerProposals = ({
               proposal_id: proposal.id,
             });
           }}
+          data-testid="proposal"
         >
           <ProposalCard proposalProps={proposal} />
         </Link>