+++ packages/playwright-config-custom/index.ts
@@ -1,0 +1,2 @@
+export { sharedConfig } from "./playwright.shared.config";
+export { helpers } from "./helpers";