+++ packages/vitest-config-custom/vitest.shared.config.ts
@@ -8,5 +8,12 @@ module.exports = {
     deps: {
       moduleDirectories: [path.resolve("../../packages"), "node_modules"],
     },
+    exclude: [
+      "**/node_modules/**",
+      "**/dist/**",
+      "**/e2e/**",
+      "**/.{idea,git,cache,output,temp}/**",
+      "**/{karma,rollup,webpack,vite,vitest,jest,ava,babel,nyc,cypress,tsup,build}.config.*",
+    ],
   },
 };